+++ libc/src/sys/mman/linux/pkey_free.cpp
@@ -25,7 +25,7 @@ LLVM_LIBC_FUNCTION(int, pkey_free, (int pkey)) {
 #else
   int ret = LIBC_NAMESPACE::syscall_impl<int>(SYS_pkey_free, pkey);
   if (ret < 0) {
-    libc_errno = static_cast<int>(-ret);
+    libc_errno = -ret;
     return -1;
   }
   return 0;
